+++ src/Pim/Bundle/CatalogBundle/Datagrid/AttributeDatagridManager.php
@@ -13,7 +13,7 @@ use Oro\Bundle\GridBundle\Property\UrlProperty;
 use Oro\Bundle\GridBundle\Property\TwigTemplateProperty;
 
 use Pim\Bundle\CatalogBundle\Manager\ProductManager;
-use Pim\Bundle\CatalogBundle\Manager\ProductAttributeManagerInterface;
+use Pim\Bundle\CatalogBundle\Manager\AttributeManagerInterface;
 
 /**
  * Product attribute grid manager
@@ -30,7 +30,7 @@ class AttributeDatagridManager extends DatagridManager
     protected $productManager;
 
     /**
-     * @var ProductAttributeManagerInterface
+     * @var AttributeManagerInterface
      */
     protected $attributeManager;
 
@@ -43,9 +43,9 @@ class AttributeDatagridManager extends DatagridManager
     }
 
     /**
-     * @param ProductAttributeManagerInterface $manager
+     * @param AttributeManagerInterface $manager
      */
-    public function setAttributeManager(ProductAttributeManagerInterface $manager)
+    public function setAttributeManager(AttributeManagerInterface $manager)
     {
         $this->attributeManager = $manager;
     }
@@ -165,7 +165,7 @@ class AttributeDatagridManager extends DatagridManager
                 'field_options' => array('choices' => $this->getAttributeTypeFieldOptions(), 'multiple' => true),
             )
         );
-        $templateProperty = new TwigTemplateProperty($field, 'PimCatalogBundle:ProductAttribute:_field-type.html.twig');
+        $templateProperty = new TwigTemplateProperty($field, 'PimCatalogBundle:Attribute:_field-type.html.twig');
         $field->setProperty($templateProperty);
 
         return $field;