+++ src/Pim/Bundle/CatalogBundle/DependencyInjection/Compiler/ResolveDoctrineOrmTargetEntitiesPass.php
@@ -17,11 +17,11 @@ class ResolveDoctrineOrmTargetEntitiesPass extends AbstractResolveDoctrineOrmTar
     protected function getParametersMapping()
     {
         return array(
-            'Symfony\Component\Security\Core\User\UserInterface'       => 'oro_user.entity.class',
-            'Pim\Bundle\CatalogBundle\Model\ProductValueInterface'     => 'pim_catalog.entity.product_value.class',
-            'Pim\Bundle\CatalogBundle\Model\ProductAttributeInterface' => 'pim_catalog.entity.product_attribute.class',
-            'Pim\Bundle\CatalogBundle\Model\ProductInterface'          => 'pim_catalog.entity.product.class',
-            'Pim\Bundle\CatalogBundle\Model\CategoryInterface'         => 'pim_catalog.entity.category.class',
+            'Symfony\Component\Security\Core\User\UserInterface'   => 'oro_user.entity.class',
+            'Pim\Bundle\CatalogBundle\Model\ProductValueInterface' => 'pim_catalog.entity.product_value.class',
+            'Pim\Bundle\CatalogBundle\Model\AttributeInterface'    => 'pim_catalog.entity.product_attribute.class',
+            'Pim\Bundle\CatalogBundle\Model\ProductInterface'      => 'pim_catalog.entity.product.class',
+            'Pim\Bundle\CatalogBundle\Model\CategoryInterface'     => 'pim_catalog.entity.category.class',
         );
     }
 }