+++ src/Pim/Bundle/CatalogBundle/Entity/Repository/AttributeRepository.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\CatalogBundle\Entity\Repository;
 
-use Pim\Bundle\FlexibleEntityBundle\Entity\Repository\AttributeRepository;
+use Pim\Bundle\FlexibleEntityBundle\Entity\Repository\AttributeRepository as FlexibleAttributeRepository;
 
 /**
  * Repository for attribute entity
@@ -11,7 +11,7 @@ use Pim\Bundle\FlexibleEntityBundle\Entity\Repository\AttributeRepository;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ProductAttributeRepository extends AttributeRepository implements ReferableEntityRepositoryInterface
+class AttributeRepository extends FlexibleAttributeRepository implements ReferableEntityRepositoryInterface
 {
     /**
      * @return \Doctrine\Common\Collections\ArrayCollection