+++ src/Pim/Bundle/CatalogBundle/Form/Type/AttributeProperty/ScopableType.php
@@ -6,7 +6,7 @@ use Symfony\Component\OptionsResolver\OptionsResolverInterface;
 use Symfony\Component\Form\AbstractType;
 
 /**
- * Form type related to scopable property of ProductAttributeInterface
+ * Form type related to scopable property of AttributeInterface
  *
  * @author    Filips Alpe <filips@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)